#175 Refactor: BookmarkService 통합
