+++ linknamu/src/main/java/com/kakao/linknamu/bookmarktag/service/BookmarkTagSaveService.java
@@ -1,10 +1,13 @@
 package com.kakao.linknamu.bookmarktag.service;
 
+import com.kakao.linknamu.bookmark.BookmarkExceptionStatus;
 import com.kakao.linknamu.bookmark.entity.Bookmark;
-import com.kakao.linknamu.bookmark.service.BookmarkReadService;
+import com.kakao.linknamu.bookmark.repository.BookmarkJpaRepository;
+import com.kakao.linknamu.bookmark.service.BookmarkService;
 import com.kakao.linknamu.bookmarktag.dto.CreateBookmarkTagRequestDto;
 import com.kakao.linknamu.bookmarktag.entity.BookmarkTag;
 import com.kakao.linknamu.bookmarktag.repository.BookmarkTagJpaRepository;
+import com.kakao.linknamu.core.exception.Exception404;
 import com.kakao.linknamu.tag.entity.Tag;
 import com.kakao.linknamu.tag.service.TagSaveService;
 import com.kakao.linknamu.tag.service.TagSearchService;
@@ -21,9 +24,9 @@ import java.util.List;
 public class BookmarkTagSaveService {
 	private final BookmarkTagJpaRepository bookmarkTagJPARepository;
 	private final BookmarkTagReadService bookmarkTagReadService;
-	private final BookmarkReadService bookmarkReadService;
 	private final TagSearchService tagSearchService;
 	private final TagSaveService tagSaveService;
+	private final BookmarkJpaRepository bookmarkJpaRepository;
 
 	public void create(List<BookmarkTag> bookmarkTagList) {
 		bookmarkTagJPARepository.saveAll(bookmarkTagList);
@@ -44,7 +47,8 @@ public class BookmarkTagSaveService {
 				return newTag;
 			});
 
-		Bookmark bookmark = bookmarkReadService.getBookmarkById(bookmarkId);
+		Bookmark bookmark = bookmarkJpaRepository.findById(bookmarkId).orElseThrow(() ->
+			new Exception404(BookmarkExceptionStatus.BOOKMARK_NOT_FOUND));
 
 		bookmarkTagReadService.validDuplicatedBookmarkTag(bookmarkId, tag.getTagId());
