+++ linknamu/src/main/java/com/kakao/linknamu/share/service/category/CreateCategoryFromEncodedIdService.java
@@ -1,8 +1,7 @@
 package com.kakao.linknamu.share.service.category;
 
 import com.kakao.linknamu.bookmark.entity.Bookmark;
-import com.kakao.linknamu.bookmark.service.BookmarkCreateService;
-import com.kakao.linknamu.bookmark.service.BookmarkReadService;
+import com.kakao.linknamu.bookmark.service.BookmarkService;
 import com.kakao.linknamu.bookmarktag.service.BookmarkTagReadService;
 import com.kakao.linknamu.category.entity.Category;
 import com.kakao.linknamu.category.service.CategoryService;
@@ -26,9 +25,8 @@ public class CreateCategoryFromEncodedIdService {
 	private final WorkspaceService workspaceService;
 	private final AesEncryption aesEncryption;
 	private final CategoryService categoryService;
-	private final BookmarkReadService bookmarkReadService;
-	private final BookmarkCreateService bookmarkCreateService;
 	private final BookmarkTagReadService bookmarkTagReadService;
+	private final BookmarkService bookmarkService;
 
 	public void createCategory(String encodedCategoryId, CreateCategoryFromEncodedIdRequestDto requestDto, User user) {
 		String categoryId = aesEncryption.decode(encodedCategoryId);
@@ -36,10 +34,10 @@ public class CreateCategoryFromEncodedIdService {
 		Category category = categoryService.findById(id);
 		Workspace workspace = workspaceService.getWorkspaceById(requestDto.workSpaceId());
 		Category newCategory = categoryService.save(category.getCategoryName(), workspace);
-		List<Bookmark> bookmarkList = bookmarkReadService.getBookmarkListByCategoryId(category.getCategoryId());
+		List<Bookmark> bookmarkList = bookmarkService.getBookmarkListByCategoryId(category.getCategoryId());
 		for (Bookmark bookmark : bookmarkList) {
 			List<Tag> tagList = bookmarkTagReadService.findTagByBookmarkId(bookmark.getBookmarkId());
-			bookmarkCreateService.addBookmark(bookmark, newCategory, tagList, user);
+			bookmarkService.addBookmark(bookmark, newCategory, tagList, user);
 		}
 
 	}
