+++ linknamu/src/main/java/com/kakao/linknamu/share/service/workspace/CreateWorkspaceFromEncodedIdService.java
@@ -1,8 +1,7 @@
 package com.kakao.linknamu.share.service.workspace;
 
 import com.kakao.linknamu.bookmark.entity.Bookmark;
-import com.kakao.linknamu.bookmark.service.BookmarkCreateService;
-import com.kakao.linknamu.bookmark.service.BookmarkReadService;
+import com.kakao.linknamu.bookmark.service.BookmarkService;
 import com.kakao.linknamu.bookmarktag.service.BookmarkTagReadService;
 import com.kakao.linknamu.category.entity.Category;
 import com.kakao.linknamu.category.service.CategoryService;
@@ -31,10 +30,9 @@ public class CreateWorkspaceFromEncodedIdService {
 	private final AesEncryption aesEncryption;
 	private final WorkspaceSaveService workspaceSaveService;
 	private final CategoryService categoryService;
-	private final BookmarkReadService bookmarkReadService;
-	private final BookmarkCreateService bookmarkCreateService;
 	private final WorkspaceReadService workspaceReadService;
 	private final BookmarkTagReadService bookmarkTagReadService;
+	private final BookmarkService bookmarkService;
 
 	public void createWorkSpace(String encodedWorkSpaceId, User user) {
 		String workspaceId = aesEncryption.decode(encodedWorkSpaceId);
@@ -50,10 +48,10 @@ public class CreateWorkspaceFromEncodedIdService {
 
 		for (Category category : workspace.getCategorySet()) {
 			Category newCategory = categoryService.save(category.getCategoryName(), newWorkspace);
-			List<Bookmark> bookmarkList = bookmarkReadService.getBookmarkListByCategoryId(category.getCategoryId());
+			List<Bookmark> bookmarkList = bookmarkService.getBookmarkListByCategoryId(category.getCategoryId());
 			for (Bookmark bookmark : bookmarkList) {
 				List<Tag> tagList = bookmarkTagReadService.findTagByBookmarkId(bookmark.getBookmarkId());
-				bookmarkCreateService.addBookmark(bookmark, newCategory, tagList, user);
+				bookmarkService.addBookmark(bookmark, newCategory, tagList, user);
 
 			}
 		}
