+++ linknamu/src/main/java/com/kakao/linknamu/workspace/controller/WorkspaceController.java
@@ -36,7 +36,6 @@ public class WorkspaceController {
 		return ResponseEntity.ok(ApiUtils.success(null));
 	}
 
-
 	@PostMapping("/update/{workspaceId}")
 	public ResponseEntity<?> updateWorkspace(
 		@PathVariable("workspaceId") Long workspaceId,
